+++ R/import.R
@@ -3,7 +3,7 @@
 #' @name MzTabM-import
 #'
 #' @description
-#' Reads and parses an mzTab-M (version 2.1) file into a named list.
+#' Reads and parses an mzTab-M (version 2.1) file into a MzTabM object.
 #'
 #' The parser recognises the four standard sections:
 #' - **MTD** (Metadata): Always required. The metadata section
@@ -31,7 +31,7 @@
 #'     `data.table::fread`.
 #'
 #' @return
-#' A named `list` containing between one and four elements, depending on
+#' A `MzTabM` object containing between one and four elements, depending on
 #' which sections are present in the file:
 #' - `MTD`: A `matrix` with two columns. Column names are absent (the
 #'     prefix column is dropped).
@@ -51,7 +51,7 @@
 #' ## Basic usage
 #' result <- readMzTabM(system.file("mztabm/out","xcms_mzTab-M_2-1_v2.mzTab",
 #'                                    package = "RmzTabM"))
-#' names(result)
+#' result
 #'
 #' @export
 readMzTabM <- function(path, ...) {
@@ -71,25 +71,25 @@ readMzTabM <- function(path, ...) {
               for SMF section, which is required if SME section is present.")
 
     ## TODO: manage the errors.
-    mzTabMValidator(path)
+    validation_res <- mzTabMValidator(path)
 
     out <- list()
     ## MTD
     mtd_lines <- grep("^MTD\\t", lines, value = TRUE)
-    out[["MTD"]] <- as.matrix(fread(text = mtd_lines, sep = "\t",
+    out[["mtd"]] <- as.matrix(fread(text = mtd_lines, sep = "\t",
                                     header = FALSE, fill = TRUE, drop = 1, ...))
 
     ## SML
     if (any(grepl("^SMH\\t", lines))) {
         sml_lines <- grep("^SMH\\t|^SML\\t", lines, value = TRUE)
-        out[["SML"]] <- as.matrix(fread(text = sml_lines, sep = "\t",
+        out[["sml"]] <- as.matrix(fread(text = sml_lines, sep = "\t",
                                         header = TRUE, fill = TRUE, ...))
     }
 
     ## SMF
     if (any(grepl("^SFH\\t", lines))) {
         smf_lines <- grep("^SFH\\t|^SMF\\t", lines, value = TRUE)
-        out[["SMF"]] <- as.matrix(fread(text = smf_lines, sep = "\t",
+        out[["smf"]] <- as.matrix(fread(text = smf_lines, sep = "\t",
                                         header = TRUE, fill = TRUE, ...))
     }
 
@@ -97,10 +97,11 @@ readMzTabM <- function(path, ...) {
     if (any(grepl("^SFH\\t", lines)) && any(grepl("^SEH\\t", lines))) {
         ## SME
         sme_lines <- grep("^SEH\\t|^SME\\t", lines, value = TRUE)
-        out[["SME"]] <- as.matrix(fread(text = sme_lines, sep = "\t",
+        out[["sme"]] <- as.matrix(fread(text = sme_lines, sep = "\t",
                                         header = TRUE, fill = TRUE, ...))
     }
 
+    out <- do.call(MzTabM, out)
     out
 }